[FE] refactor: AddGuildButton 컴포넌트 병합 (#31)
